drivers/rtc/rtc-twl.c: ensure all interrupts are disabled during probe
On some platforms, bootloaders are known to do some interesting RTC
programming.  Without going into the obscurities as to why this may be the
case, suffice it to say the the driver should not make any assumptions
about the state of the RTC when the driver loads.  In particular, the
driver probe should be sure that all interrupts are disabled until
otherwise programmed.

This was discovered when finding bursty I2C traffic every second on Overo
platforms.  This I2C overhead was keeping the SoC from hitting deep power
states.  The cause was found to be the RTC firing every second on the
I2C-connected TWL PMIC.

Special thanks to Felipe Balbi for suggesting to look for a rogue driver
as the source of the I2C traffic rather than the I2C driver itself.

Special thanks to Steve Sakoman for helping track down the source of the
continuous RTC interrups on the Overo boards.

Signed-off-by: Kevin Hilman <khilman@ti.com>
Cc: Felipe Balbi <balbi@ti.com>
Cc: Steve Sakoman <steve@sakoman.com>
Cc: Alessandro Zummo <a.zummo@towertech.it>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>