Enable wilink8 wifi chip for Overo
A Wilink WL18xx chip can replace either the W2CBW003 or W2CBW0015
wifi + bluetooth modules found on some Gumstix Overo COMs.  Add
platform data to support this chip if it is present.  Here, board
configuration data is encapsulated in pdata-quirks.c rather than via
a device tree file as the DT-bindings for this chip are not yet in
mainline kernels.
Note: Bluetooth is not yet supported

Signed-off-by: Ash Charles <ashcharles@gmail.com>