Set status of MCBSP2 for audio on Overo Storm
Gumstix Overo and Overo Storm COMs use TWL4030 audio module. The module is
connected to the 2nd module of MCBSP on the OMAP3 SoC. The status is set to
"okay" only on omap3-overo.dtsi, and not on omap3-overo-storm.dtsi. Unless it
is specified, it is assumed to be disabled. This resulted in broken audio on
Storm COMs.

Signed-off-by: Ash Charles <ashcharles@gmail.com>