Enable wilink8 bluetooth chip for Overo
Add platform data for Gumstix Overo COMs to support bluetooth from the
Wilink WL18xx chips.  Here, board configuration data is encapsulated
in pdata-quirks.c rather than via device tree file as the DT-bindings
for this bluetooth modules are not yes in mainline kernels. We need to
disable the GPIO nReset for other bluetooth modules as the btwilink
wants to claim it.

Signed-off-by: Ash Charles <ashcharles@gmail.com>