arm: Add Gumstix Pepper43R and Pepper43C boards
Add device tree supporting the Gumstix Pepper43C [1] and Pepper43R [2]
boards.  These feature, respectively, a 4.3" capacitive or resistive
touchscreen but are otherwise mutually identical and similar to the
original, deprecated, Gumstix Pepper.

This commit could be improved by creating a common dtsi for the
shared aspects of these boards.

[1] https://store.gumstix.com/computers/pepper-43c.html
[2] https://store.gumstix.com/computers/pepper-43r.html

Signed-off-by: Ash Charles <ashcharles@gmail.com>