omap3-overo: Update device tree for Arbor expansion boards
Make a few pin allocation tweaks to support the production rev. of
the Arbor43C (v1), Arbor50C and Arbor70C boards.  IMU chips are added
on a SPI bus and this is reflected in the device trees however bindings
for the driver are still needed.

Additionally, correct screen panel timings that were erroneously
altered in omap3-overo-common-lcd43 rather than on a more
screen-specific level.

Signed-off-by: Ash Charles <ashcharles@gmail.com>