tfp410: RGB888 not RGB565 color-mapping for botball
Botball wires 24-bit color to the TFP410 HDMI encoder---the driver
should reflect this.  Normally, we'd just lose a little colo depth but
because of the AM335x uses a strange pin-mapping for RGB888 compared
with RGB565, we actually end up partially switching the blue and red
channels if the color-depth is set wrongly.

This is should be fixed properly by making the 'bpp' parameter settable
from the device-tree for the tfp410.

Signed-off-by: Ash Charles <ashcharles@gmail.com>