am437x-poblano: Add support for accel, magn, gyro
LSM303D - Currently there is no mainline driver for
this sensor. So download the driver from ST website:
http://www.st.com/web/catalog/tools/FM147/CL1818/SC1885/PF258118
Modify it to be device tree compatible and add it as
an input miscelleanous driver. Enable it as
CONFIG_INPUT_LSM303D in the kernel config.

L3GD20H - This is listed as one of the supported devices
under the iio subsystem but the WHO_AM_I register value
is listed as 0xd4 instead of 0xd7.

Signed-off-by: Arun Bharadwaj <arun@gumstix.com>