edt-ft5x06: Handle reset/wake for different versions of Newheaven display
In the driver, the function devm_gpiod_get_optional parses the wake/reset - gpios in the device tree.
If the device tree does not specify reset-gpios, the function will return NULL.
treat the reset as wake by using the same function but with differnt parameters.

The reset-gpios in the driver will obtain the content of the wake-gpios which is specified in the device tree.
Set wake-gpio in the driver to NULL in order to avoid some issues related to undelcared variable.

Removing all the codes related to wake-gpios becuase it is not needed anymore.

Signed-off-by: Jerry Hung <jerry@gumstix.com>